updated windows-rs version, updated other libs to latest patch

+++ crates/windows-capture-provider/src/display_cache/mod.rs
@@ -32,7 +32,8 @@ impl DisplayCache {
                 .create_capture_item()
                 .map_err(Error::CreateCaputreItem)?;
 
-            capture_items.insert(display.handle.0, capture_item);
+            // TODO ensure correct key
+            capture_items.insert(display.handle.0 as isize, capture_item);
         }
 
         Ok(Self {
